% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{launch_workshop}
\alias{launch_workshop}
\title{Lance l'atelier}
\usage{
launch_workshop(file, port = 8181, host = "localhost")
}
\arguments{
\item{file}{le file Rmd du tutoriel}

\item{port}{le port sur lequel on envoie la page web}

\item{host}{l'url sur laquelle on publie}
}
\description{
Lance l'atelier
}
